Simplify Variation results
