+++ internal/model/variation_result.go
@@ -2,60 +2,30 @@ package model
 
 import "github.com/thomaspoignant/go-feature-flag/internal/flag"
 
-type VariationResult struct {
+// JSONType contains all acceptable flag value types
+type JSONType interface {
+	float64 | int | string | bool | interface{} | map[string]interface{}
+}
+
+// VariationResult contains all the field available in a flag variation result.
+type VariationResult[T JSONType] struct {
 	TrackEvents   bool                  `json:"trackEvents"`
 	VariationType string                `json:"variationType"`
 	Failed        bool                  `json:"failed"`
 	Version       string                `json:"version"`
 	Reason        flag.ResolutionReason `json:"reason"`
 	ErrorCode     flag.ErrorCode        `json:"errorCode"`
-}
-
-// BoolVarResult is the internal result format of a bool variation.
-// This is used by the ffclient.BoolVariation functions and by the ffclient.AllFlagsState function
-type BoolVarResult struct {
-	VariationResult
-	Value bool
-}
-
-// IntVarResult is the internal result format of a int variation.
-// This is used by the ffclient.IntVariation functions and by the ffclient.AllFlagsState function
-type IntVarResult struct {
-	VariationResult
-	Value int
-}
-
-// Float64VarResult is the internal result format of a float64 variation.
-// This is used by the ffclient.Float64Variation functions and by the ffclient.AllFlagsState function
-type Float64VarResult struct {
-	VariationResult
-	Value float64
-}
-
-// StringVarResult is the internal result format of a string variation.
-// This is used by the ffclient.StringVariation functions and by the ffclient.AllFlagsState function
-type StringVarResult struct {
-	VariationResult
-	Value string
-}
-
-// JSONVarResult is the internal result format of a json variation.
-// This is used by the ffclient.JSONVariation functions and by the ffclient.AllFlagsState function
-type JSONVarResult struct {
-	VariationResult
-	Value map[string]interface{}
-}
-
-// JSONArrayVarResult is the internal result format of a json array variation.
-// This is used by the ffclient.JSONArrayVariation functions and by the ffclient.AllFlagsState function
-type JSONArrayVarResult struct {
-	VariationResult
-	Value []interface{} `json:"value"`
+	Value         T                     `json:"value"`
 }
 
 // RawVarResult is the result of the raw variation call.
 // This is used by ffclient.RawVariation functions, this should be used only by internal calls.
 type RawVarResult struct {
-	VariationResult
-	Value interface{} `json:"value"`
+	TrackEvents   bool                  `json:"trackEvents"`
+	VariationType string                `json:"variationType"`
+	Failed        bool                  `json:"failed"`
+	Version       string                `json:"version"`
+	Reason        flag.ResolutionReason `json:"reason"`
+	ErrorCode     flag.ErrorCode        `json:"errorCode"`
+	Value         interface{}           `json:"value"`
 }
