+++ variation_test.go
@@ -1918,13 +1918,11 @@ func TestRawVariation(t *testing.T) {
 			},
 			wantErr: true,
 			want: model.RawVarResult{
-				VariationResult: model.VariationResult{
-					VariationType: flag.VariationSDKDefault,
-					Failed:        true,
-					Reason:        flag.ReasonError,
-					ErrorCode:     flag.ErrorCodeProviderNotReady,
-				},
-				Value: "",
+				VariationType: flag.VariationSDKDefault,
+				Failed:        true,
+				Reason:        flag.ReasonError,
+				ErrorCode:     flag.ErrorCodeProviderNotReady,
+				Value:         "",
 			},
 		},
 		{
@@ -1938,13 +1936,11 @@ func TestRawVariation(t *testing.T) {
 				}, nil),
 			},
 			want: model.RawVarResult{
-				Value: true,
-				VariationResult: model.VariationResult{
-					VariationType: flag.VariationSDKDefault,
-					Failed:        false,
-					TrackEvents:   true,
-					Reason:        flag.ReasonDisabled,
-				},
+				Value:         true,
+				VariationType: flag.VariationSDKDefault,
+				Failed:        false,
+				TrackEvents:   true,
+				Reason:        flag.ReasonDisabled,
 			},
 			wantErr:     false,
 			expectedLog: "^\\[" + testutils.RFC3339Regex + "\\] user=\"random-key\", flag=\"disable-flag\", value=\"true\", variation=\"SdkDefault\"\n",
@@ -1960,14 +1956,12 @@ func TestRawVariation(t *testing.T) {
 					errors.New("impossible to read the toggle before the initialisation")),
 			},
 			want: model.RawVarResult{
-				Value: "defaultValue",
-				VariationResult: model.VariationResult{
-					VariationType: flag.VariationSDKDefault,
-					Failed:        true,
-					TrackEvents:   true,
-					Reason:        flag.ReasonError,
-					ErrorCode:     flag.ErrorCodeFlagNotFound,
-				},
+				Value:         "defaultValue",
+				VariationType: flag.VariationSDKDefault,
+				Failed:        true,
+				TrackEvents:   true,
+				Reason:        flag.ReasonError,
+				ErrorCode:     flag.ErrorCodeFlagNotFound,
 			},
 			wantErr:     true,
 			expectedLog: "^\\[" + testutils.RFC3339Regex + "\\] user=\"random-key\", flag=\"key-not-exist\", value=\"defaultValue\", variation=\"SdkDefault\"\n",
@@ -1981,14 +1975,12 @@ func TestRawVariation(t *testing.T) {
 				cacheMock:    NewCacheMock(&flag.InternalFlag{}, errors.New("flag [key-not-exist] does not exists")),
 			},
 			want: model.RawVarResult{
-				Value: 123456,
-				VariationResult: model.VariationResult{
-					VariationType: flag.VariationSDKDefault,
-					Failed:        true,
-					TrackEvents:   true,
-					Reason:        flag.ReasonError,
-					ErrorCode:     flag.ErrorCodeFlagNotFound,
-				},
+				Value:         123456,
+				VariationType: flag.VariationSDKDefault,
+				Failed:        true,
+				TrackEvents:   true,
+				Reason:        flag.ReasonError,
+				ErrorCode:     flag.ErrorCodeFlagNotFound,
 			},
 			wantErr:     true,
 			expectedLog: "^\\[" + testutils.RFC3339Regex + "\\] user=\"random-key\", flag=\"key-not-exist\", value=\"123456\", variation=\"SdkDefault\"\n",
@@ -2022,13 +2014,11 @@ func TestRawVariation(t *testing.T) {
 				}, nil),
 			},
 			want: model.RawVarResult{
-				Value: map[string]interface{}{"test": "test"},
-				VariationResult: model.VariationResult{
-					VariationType: "Default",
-					Failed:        false,
-					TrackEvents:   true,
-					Reason:        flag.ReasonDefault,
-				},
+				Value:         map[string]interface{}{"test": "test"},
+				VariationType: "Default",
+				Failed:        false,
+				TrackEvents:   true,
+				Reason:        flag.ReasonDefault,
 			},
 			wantErr:     false,
 			expectedLog: "^\\[" + testutils.RFC3339Regex + "\\] user=\"random-key\", flag=\"test-flag\", value=\"map\\[test:test\\]\", variation=\"Default\"",
@@ -2062,13 +2052,11 @@ func TestRawVariation(t *testing.T) {
 				}, nil),
 			},
 			want: model.RawVarResult{
-				Value: map[string]interface{}{"test2": "test"},
-				VariationResult: model.VariationResult{
-					VariationType: "True",
-					Failed:        false,
-					TrackEvents:   true,
-					Reason:        flag.ReasonTargetingMatch,
-				},
+				Value:         map[string]interface{}{"test2": "test"},
+				VariationType: "True",
+				Failed:        false,
+				TrackEvents:   true,
+				Reason:        flag.ReasonTargetingMatch,
 			},
 			wantErr:     false,
 			expectedLog: "^\\[" + testutils.RFC3339Regex + "\\] user=\"random-key\", flag=\"test-flag\", value=\"map\\[test2:test\\]\", variation=\"True\"",
@@ -2102,13 +2090,11 @@ func TestRawVariation(t *testing.T) {
 				}, nil),
 			},
 			want: model.RawVarResult{
-				Value: map[string]interface{}{"test3": "test"},
-				VariationResult: model.VariationResult{
-					VariationType: "False",
-					Failed:        false,
-					TrackEvents:   true,
-					Reason:        flag.ReasonTargetingMatch,
-				},
+				Value:         map[string]interface{}{"test3": "test"},
+				VariationType: "False",
+				Failed:        false,
+				TrackEvents:   true,
+				Reason:        flag.ReasonTargetingMatch,
 			},
 			wantErr:     false,
 			expectedLog: "^\\[" + testutils.RFC3339Regex + "\\] user=\"random-key-ssss1\", flag=\"test-flag\", value=\"map\\[test3:test\\]\", variation=\"False\"",
@@ -2143,13 +2129,11 @@ func TestRawVariation(t *testing.T) {
 				}, nil),
 			},
 			want: model.RawVarResult{
-				Value: true,
-				VariationResult: model.VariationResult{
-					VariationType: "True",
-					Failed:        false,
-					TrackEvents:   false,
-					Reason:        flag.ReasonTargetingMatch,
-				},
+				Value:         true,
+				VariationType: "True",
+				Failed:        false,
+				TrackEvents:   false,
+				Reason:        flag.ReasonTargetingMatch,
 			},
 			wantErr:     false,
 			expectedLog: "^$",
@@ -2166,12 +2150,10 @@ func TestRawVariation(t *testing.T) {
 				}, nil),
 			},
 			want: model.RawVarResult{
-				Value: false,
-				VariationResult: model.VariationResult{
-					VariationType: flag.VariationSDKDefault,
-					Failed:        false,
-					TrackEvents:   false,
-				},
+				Value:         false,
+				VariationType: flag.VariationSDKDefault,
+				Failed:        false,
+				TrackEvents:   false,
 			},
 			wantErr:     false,
 			expectedLog: "",
